+++ src/app/mentorship/page.tsx
@@ -5,7 +5,7 @@ import { pageStore } from '@/entities/page/model/store';
 import { mentorshipMetadata } from '@/metadata/mentorship';
 import { OG_SITE_NAME } from '@/shared/constants';
 import { generatePageMetadata } from '@/shared/helpers/generate-page-metadata';
-import { Mentorship } from '@/views/mentorship/mentorship';
+import { Mentorship } from '@/views/mentorship';
 
 export async function generateMetadata(): Promise<Metadata> {
   const {
